+++ src/components/InternalLink.tsx
@@ -1,12 +1,8 @@
 import React, { FunctionComponent } from 'react';
 import Link from 'next/link';
+import { LinkProps } from './Link';
 
-interface InternalLinkProps {
-  href: string;
-  className?: string;
-}
-
-const InternalLink: FunctionComponent<InternalLinkProps> = ({ children, href, className, ...rest }) => (
+const InternalLink: FunctionComponent<LinkProps> = ({ children, href, className, ...rest }) => (
   <Link href={href} {...rest}>
     <a className={className}>{children}</a>
   </Link>
